OKAY, tabler.range NOW is fully completegit add .!
I drastically  improved the code efficiency when using the center argument - now, new start and stop values are calculated with math.floor, and a single call to algo is made. Before, we had to do manual array concatenation and reversal. --- ALSO, strange things happen when FP_ERROR is set to 0 (In tabler.range). See comments.